Replace CWMSpy with cwms-python v0.3.0
Significant changes to process_write_tasks() were required due to a
change in the return value of write_ts()/store_timeseries().  write_ts()
returned the response, whereas store_timeseries() returns None and
raises an error if the request fails.  Because the initial request
payload can't be retrieved from the return value it is now stored
separately in self._parsed_payloads and retrieved using the index of
enumerate(asyncio.gather).